feat: implement union type generation rules and fix type mapping
- Implement union type strategy in CJTypeMapper (overload sentinel/field first-type retention)
- Add default Float64 strategy in CJNumberConversion for ambiguous number types
- Implement parameter overload logic in ComponentsPrinter and PeersPrinter
- Disable Union_*/Tuple_* type names in generated code
- Fix ResourceStr overload logic and IDL type representation

Results:
Union type overloads correctly generated
Union_* and Tuple_* type names eliminated
Reasonable Any type usage

+++ arkgen/src/declaration/CJNumberConversion.ts
@@ -100,8 +100,8 @@ export function mapNumberType(context: NumberMappingContext): 'Length' | 'Float6
     return 'Int64';
   }
   
-  // 默认场景: 使用 Int64 类型
-  return 'Int64';
+  // 默认场景: 使用 Float64 类型（无法判定语义时统一使用 Float64）
+  return 'Float64';
 }
 
 /**
